Car dealer home pages
My first ever web page created on my fundation year at Staffordshire University on 01 January 2019.  Ihad to use a bit of my knowledge on illustrator to coe up with a original log type for the businessand treatment of images on photoshop, during that period got myself my first expose to css and a good practise how to work both css and html.